ERF: Move testing utilities to make them accessible from doctests
